link users to todos

+++ convex/functions.ts
@@ -1,9 +1,14 @@
 import { v } from "convex/values";
 import {query, mutation} from "./_generated/server";
+import { requireUser } from "./helpers";
 
 export const listTodos = query({
     handler: async (ctx) => {
-        return await ctx.db.query("todos").collect();
+        const user = await requireUser(ctx);
+        return await ctx.db.query("todos")
+        .withIndex("by_user_id",q => q.eq("userId",user.tokenIdentifier))
+        // .filter(q => q.eq(q.field("userId"),user.tokenIdentifier))
+        .collect();
     } 
 })
 
@@ -13,10 +18,12 @@ export const CreateTodo = mutation({
         description: v.string(),        
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
         await ctx.db.insert("todos",{
             title: args.title,
             description: args.description,
-            completed: false
+            completed: false,
+            userId:user.tokenIdentifier
         })
     }
 });
@@ -27,6 +34,11 @@ export const updateTodo = mutation({
         completed: v.boolean()
     },
     handler: async (ctx,args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id);
+        if (todo?.userId !== user.tokenIdentifier) {
+            throw new Error("Unauthorized");
+        }
         await ctx.db.patch(args.id, {
             completed: args.completed
         })
@@ -38,6 +50,11 @@ export const deleteTodo = mutation({
         id: v.id("todos")
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id);
+        if (todo?.userId !== user.tokenIdentifier) {
+            throw new Error("Unauthorized");
+        }
         await ctx.db.delete(args.id);
     }
 })
